adjusted to be able to add wildcard textures (*.png)

+++ abmatt/build/build.py
@@ -18,11 +18,14 @@ def main(args):
     bit_width = c['bit_width']
     version = c['version']
     interpreter = c['64-bit'] if bit_width == 'x64' else c['32-bit']
+    name = c['build_name']
+    build_type = c['build_type']
     # build
-    out_file, is_dir = build(interpreter)
+    clean(name, [name + '.exe', name])
+    my_platform = platform.system().lower()
+    out_file, is_dir = build(name, build_type, interpreter, my_platform)
     if not out_file:
         sys.exit(1)
-    my_platform = platform.system().lower()
     dist_dir = 'abmatt_' + my_platform + '-' + platform.release() + '_' \
                + bit_width + '-' + version
     # clean
@@ -52,57 +55,65 @@ def make_distribution(dir, platform, binary_path, binary_path_is_dir):
     etc = dir + '/etc/abmatt'
     shutil.copy('../presets.txt', etc)
     shutil.copy('../config.conf', etc)
+    dest_dir = os.path.join(dir, 'bin')
+    bin_dir, base_name = os.path.split(binary_path)
+    exe = os.path.join(dest_dir, 'abmatt')
+    if 'exe' in base_name:
+        exe += '.exe'
     if binary_path_is_dir:
-        bin_dir, base_name = os.path.split(binary_path)
-        shutil.copytree(bin_dir, dir + '/bin')
+        shutil.copytree(bin_dir, dest_dir)
+        shutil.move(os.path.join(dest_dir, base_name), exe)
     else:
-        dest_dir = os.path.join(dir, 'bin')
         os.mkdir(dest_dir)
-        shutil.copy(binary_path, dest_dir)
+        shutil.copy(binary_path, exe)
     # platform specific files
     if platform == 'windows':
-        shutil.copy('install_win.txt', dir)
-        shutil.copy('make_installer.nsi', dir)
+        shutil.copy('./install-win.txt', dir)
+        shutil.copy('./make_installer.nsi', dir)
         os.chdir(dir)
         if os.system('makensis make_installer.nsi'):
             return False
-        os.remove('make_installer.nsi')
+        os.remove('./make_installer.nsi')
         shutil.rmtree('etc')
         shutil.rmtree('bin')
+        os.chdir('..')
         if not zip(dir):
             return False
     else:
-        shutil.copy('install.sh', dir)
-        shutil.copy('uninstall.sh', dir)
-        shutil.copy('install-ubu.txt', os.path.join(dir, 'install.txt'))
+        shutil.copy('./install.sh', dir)
+        shutil.copy('./uninstall.sh', dir)
+        shutil.copy('./install-ubu.txt', os.path.join(dir, 'install.txt'))
+        os.chdir('..')
         if not tar(dir):
             return False
     return True
 
 
 def clean(folder, files):
-    if os.path.exists(folder):
+    if os.path.exists(folder) and os.path.isdir(folder):
         shutil.rmtree(folder)
     for x in files:
-        if os.path.exists(x):
+        if os.path.exists(x) and os.path.isfile(x):
             os.remove(x)
 
 
-def build(interpreter):
+def build(name, build_type, interpreter, platform):
     output = None
     os.chdir('..')
     for x in os.listdir():
         if x.endswith('.pyc'):
             os.remove(x)
-    name = 'main'
-    output_type = '--onedir'
+    if build_type == 'auto':
+        output_type = '--onedir' if platform == 'windows' else 'onefile'
+    else:
+        output_type = '--' + build_type
     is_dir = True if 'onedir' in output_type else False
     params = '-y __main__.py -p ../../venv/Lib/site-packages --name ' + name + ' ' + output_type
     result = os.system(interpreter + ' -m PyInstaller ' + params)
     os.chdir('dist')
     if not result:
         output = name if not is_dir else name + '/' + name
-        if not os.path.exists(output):
+        if not os.path.exists(output) or os.path.isdir(output):
             output += '.exe'
             if not os.path.exists(output):
                 print('Unable to find PyInstaller output file!')
